JAVA-2867: Revisit compressor substitutions
