--- core/src/main/java/com/datastax/oss/driver/internal/core/protocol/Lz4Missing.java
@@ -1,33 +1,0 @@
-/*
- * Copyright DataStax, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.datastax.oss.driver.internal.core.protocol;
-
-import java.util.function.BooleanSupplier;
-
-public class Lz4Missing implements BooleanSupplier {
-
-  private static final String LZ4_CLZ_NAME = "net.jpountz.lz4.LZ4Compressor";
-
-  @Override
-  public boolean getAsBoolean() {
-    try {
-      Class.forName(LZ4_CLZ_NAME);
-      return false;
-    } catch (ClassNotFoundException e) {
-      return true;
-    }
-  }
-}
